requirements.txt
# =============================================================================
# DSH ETL Search AI - Requirements
# =============================================================================
# Using sentence-transformers for free local embeddings + MongoDB via motor
# =============================================================================

# -----------------------------------------------------------------------------
# Core
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# -----------------------------------------------------------------------------
# HTTP & Async
# -----------------------------------------------------------------------------
httpx>=0.25.0               # Async HTTP client
aiohttp>=3.9.0              # Async HTTP client (used in http.py)
aiofiles>=23.0.0            # Async file I/O

# -----------------------------------------------------------------------------
# XML/JSON Parsing
# -----------------------------------------------------------------------------
lxml>=4.9.0
rdflib>=7.0.0

# -----------------------------------------------------------------------------
# Database (MongoDB)
# -----------------------------------------------------------------------------
motor>=3.3.0                # Async MongoDB driver
pymongo>=4.6.0              # MongoDB driver

# -----------------------------------------------------------------------------
# Embeddings (local, free)
# -----------------------------------------------------------------------------
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# -----------------------------------------------------------------------------
# PDF Processing
# -----------------------------------------------------------------------------
pypdf>=3.0.0
pdfplumber>=0.10.0              # PDF text extraction with layout

# -----------------------------------------------------------------------------
# Authentication
# -----------------------------------------------------------------------------
passlib[bcrypt]>=1.7.4          # Password hashing
bcrypt==4.0.1                   # Pinned for passlib compatibility
python-jose[cryptography]>=3.3.0  # JWT tokens

# -----------------------------------------------------------------------------
# ML / NLP
# -----------------------------------------------------------------------------
scikit-learn>=1.3.0             # TF-IDF keyword extraction fallback

# -----------------------------------------------------------------------------
# API Framework
# -----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
email-validator>=2.0.0
python-multipart>=0.0.6      # File uploads

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# -----------------------------------------------------------------------------
# Logging
# -----------------------------------------------------------------------------
structlog>=24.0.0

# -----------------------------------------------------------------------------
# Development
# -----------------------------------------------------------------------------
black>=23.0.0
isort>=5.12.0
